+++ app/src/main/java/de/k3b/android/androFotoFinder/queries/FotoSql.java
@@ -32,7 +32,9 @@ import org.osmdroid.util.GeoPoint;
 
 import java.util.ArrayList;
 import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import de.k3b.android.androFotoFinder.Global;
 import de.k3b.android.androFotoFinder.R;
@@ -386,7 +388,7 @@ public class FotoSql {
                 return c.getString(c.getColumnIndex(FotoSql.SQL_COL_PATH));
             }
         } catch (Exception ex) {
-            Log.e(Global.LOG_CONTEXT, "Cannot get path from " + uri, ex);
+            Log.e(Global.LOG_CONTEXT, "FotoSql.execGetFotoPath() Cannot get path from " + uri, ex);
         } finally {
             if (c != null) c.close();
         }
@@ -399,14 +401,18 @@ public class FotoSql {
 
         Cursor c = null;
         try {
-            c = resolver.query(SQL_TABLE_EXTERNAL_CONTENT_URI, new String[] {FotoSql.SQL_COL_PATH}, FotoSql.SQL_COL_PATH + " like ?", new String[] {pathFilter}, FotoSql.SQL_COL_PATH);
+            c = resolver.query(SQL_TABLE_EXTERNAL_CONTENT_URI, new String[]{FotoSql.SQL_COL_PATH}, FotoSql.SQL_COL_PATH + " like ?", new String[]{pathFilter}, FotoSql.SQL_COL_PATH);
             while (c.moveToNext()) {
                 result.add(c.getString(0));
             }
         } catch (Exception ex) {
-            Log.e(Global.LOG_CONTEXT, "Cannot get path from: " + FotoSql.SQL_COL_PATH + " like '" + pathFilter +"'", ex);
+            Log.e(Global.LOG_CONTEXT, "FotoSql.execGetFotoPaths() Cannot get path from: " + FotoSql.SQL_COL_PATH + " like '" + pathFilter +"'", ex);
         } finally {
             if (c != null) c.close();
+        }
+
+        if (Global.debugEnabled) {
+            Log.d(Global.LOG_CONTEXT, "FotoSql.execGetFotoPaths() result count=" + result.size());
         }
         return result;
     }
@@ -475,12 +481,13 @@ public class FotoSql {
                 return result;
             }
         } catch (Exception ex) {
-            Log.e(Global.LOG_CONTEXT, "execGetGeoRectangle: error executing " + query, ex);
+            Log.e(Global.LOG_CONTEXT, "FotoSql.execGetGeoRectangle(): error executing " + query, ex);
         } finally {
             if (c != null) c.close();
         }
         return null;
     }
+
 
     public static IGeoPoint execGetPosition(Context context, int id) {
         QueryParameterParcelable query = (QueryParameterParcelable) new QueryParameterParcelable()
@@ -499,11 +506,59 @@ public class FotoSql {
                 return result;
             }
         } catch (Exception ex) {
-            Log.e(Global.LOG_CONTEXT, "execGetGeoRectangle: error executing " + query, ex);
+            Log.e(Global.LOG_CONTEXT, "FotoSql.execGetPosition: error executing " + query, ex);
         } finally {
             if (c != null) c.close();
         }
         return null;
+    }
+
+    public static Map<String, Integer> execGetPathIdMap(Context context, String... fileNames) {
+        Map<String, Integer> result = new HashMap<String, Integer>();
+
+        if (fileNames != null) {
+            StringBuilder filter = new StringBuilder();
+            filter.append(SQL_COL_PATH).append(" in (");
+
+            int count = 0;
+            for (String fileName : fileNames) {
+                if (fileName != null) {
+                    if (count > 0) filter.append(", ");
+                    filter.append("'").append(fileName).append("'");
+                    count++;
+                }
+            }
+
+            if (count == 0) return result;
+            filter.append(")");
+
+            QueryParameterParcelable query = (QueryParameterParcelable) new QueryParameterParcelable()
+                    .setID(QUERY_TYPE_UNDEFINED)
+                    .addColumn(SQL_COL_PK, SQL_COL_PATH)
+                    .addFrom(SQL_TABLE_EXTERNAL_CONTENT_URI.toString())
+                    .addWhere(filter.toString());
+
+            Cursor c = null;
+            try {
+                c = query(context, query);
+                while (c.moveToNext()) {
+                    result.put(c.getString(1),c.getInt(0));
+                }
+            } catch (Exception ex) {
+                Log.e(Global.LOG_CONTEXT, "FotoSql.execGetPathIdMap: error executing " + query, ex);
+            } finally {
+                if (c != null) c.close();
+            }
+        }
+        return result;
+    }
+
+    public static int execUpdate(Context context, Integer id, ContentValues values) {
+        return context.getContentResolver().update(FotoSql.SQL_TABLE_EXTERNAL_CONTENT_URI, values, SQL_COL_PK + " = ?", new String[]{id.toString()});
+    }
+
+    public static Uri execInsert(Context context, ContentValues values) {
+        return context.getContentResolver().insert(FotoSql.SQL_TABLE_EXTERNAL_CONTENT_URI, values);
     }
 }